+++ image/service-available/:apache2/startup.sh
@@ -8,5 +8,3 @@ log-helper level eq trace && set -x
 ln -sf ${CONTAINER_SERVICE_DIR}/:apache2/assets/apache2.conf /etc/apache2/apache2.conf
 ln -sf ${CONTAINER_SERVICE_DIR}/:apache2/assets/conf-available/* /etc/apache2/conf-available/
 ln -sf ${CONTAINER_SERVICE_DIR}/:apache2/assets/sites-available/* /etc/apache2/sites-available/
-
-exit 0