+++ image/service-available/:nginx/startup.sh
@@ -6,5 +6,3 @@ log-helper level eq trace && set -x
 
 ln -sf ${CONTAINER_SERVICE_DIR}/:nginx/assets/nginx.conf /etc/nginx/nginx.conf
 ln -sf ${CONTAINER_SERVICE_DIR}/:nginx/assets/conf-available/* /etc/nginx/conf-available
-
-exit 0